move EvaluationResult and EvaluationResultKind to separate file
