+++ app/domain/functions/math.ts
@@ -1,4 +1,4 @@
-import { type EvaluationResult, EvaluationResultKind } from '~/domain/evaluation'
+import { type EvaluationResult, EvaluationResultKind } from '~/domain/evaluation-result'
 
 export function PI(extra?: EvaluationResult): EvaluationResult {
   if (extra) {
