+++ app/domain/functions/text.ts
@@ -1,8 +1,5 @@
-import {
-  type EvaluationResult,
-  EvaluationResultKind,
-  printEvaluationResult,
-} from '~/domain/evaluation'
+import { type EvaluationResult, EvaluationResultKind } from '~/domain/evaluation-result'
+import { printEvaluationResult } from '~/domain/evaluation'
 
 export function CONCAT(...args: EvaluationResult[]): EvaluationResult {
   let out = ''
