+++ app/domain/spreadsheet.ts
@@ -1,9 +1,6 @@
 import { type AST, AstKind } from '~/domain/ast'
-import {
-  type EvaluationResult,
-  EvaluationResultKind,
-  evaluateExpression,
-} from '~/domain/evaluation'
+import { evaluateExpression } from '~/domain/evaluation'
+import { type EvaluationResult, EvaluationResultKind } from '~/domain/evaluation-result'
 import { parse } from '~/domain/expression'
 import * as functions from '~/domain/functions'
 import { tokenize } from '~/domain/tokenizer'
